backend/requirements.txt
# Django & DRF
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
drf-yasg==1.21.7

# Database
psycopg2-binary==2.9.9
redis==5.0.1
messaging
# Channels & WebSockets
channels==3.0.5
channels-redis==3.4.1
daphne==3.0.2

# Media Processing
Pillow==10.2.0
pydub==0.25.1
python-magic==0.4.27
numpy==1.26.3

# Storage
django-storages==1.14.2
boto3==1.34.14

# Authentication
social-auth-app-django==5.4.0
python-jose==3.3.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Utils
python-dotenv==1.0.0
celery==5.3.6
flower==2.0.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-asyncio==0.23.3
coverage==7.4.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0

# Documentation
Sphinx==6.2.1
sphinx-rtd-theme==1.2.0
docutils==0.18.1

# Production
gunicorn==21.2.0
uvicorn==0.25.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.1
django-debug-toolbar==4.2.0

# Security
django-ratelimit==4.1.0
django-defender==0.9.7

# Cache
django-cacheops==7.0.2

# Search
django-elasticsearch-dsl==7.3.0

# Performance
django-silk==5.0.4

# Add these if not already present
psycopg2-binary>=2.9.1
django-postgres-extensions>=0.9.3 
drf-nested-routers